requirements.txt
asgiref==3.8.1
attrs==25.3.0
Automat==24.8.1
black==24.10.0
buildtools==1.0.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
docopt==0.6.2
furl==2.1.4
greenlet==3.1.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysqlclient==2.2.7
orderedmultidict==1.0.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.10.1
python-arango==8.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redo==3.0.0
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.8.4
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
zipp==3.21.0
zope.interface==7.2
